+++ exercises/static/exercises/follow_person/react-components/SpecificFollowPerson.js
@@ -1,9 +1,9 @@
-import * as React from "react";
-import PropTypes from "prop-types";
+import React, { useEffect, useState } from "react";
 import noImage from "../../assets/img/noImage.png";
 import "./css/GUICanvas.css";
+import PropTypes from "prop-types";
 
-export default function ImgCanvas() {
+function SpecificFollowPerson(props) {
   const [image, setImage] = React.useState(
     noImage
   );
@@ -36,13 +36,38 @@ export default function ImgCanvas() {
       );
     };
   }, []);
+
+  
+
+  useEffect(() => {
+    listen_key();
+  }, []);
+
+  function listen_key() {
+    window.addEventListener("keypress", function (event) {
+      if (event.code === "KeyS") {
+        window.RoboticsExerciseComponents.commsManager.send("gui", "key_s");
+      } else if (event.code === "KeyW") {
+        window.RoboticsExerciseComponents.commsManager.send("gui", "key_w");
+      } else if (event.code === "KeyA") {
+        window.RoboticsExerciseComponents.commsManager.send("gui", "key_a");
+      } else if (event.code === "KeyD") {
+        window.RoboticsExerciseComponents.commsManager.send("gui", "key_d");
+      } else if (event.code === "KeyX") {
+        window.RoboticsExerciseComponents.commsManager.send("gui", "key_x");
+      }
+    });
+  }
+
   return (
     <div style={{display: "flex", width: "100%", height: "100%", position:"relative", justifyContent: "center"}}>
       <img className="image" id="gui_canvas" src={image}/>
     </div>
   );
+
 }
 
-ImgCanvas.propTypes = {
+SpecificFollowPerson.propTypes = {
   context: PropTypes.any,
 };
+export default SpecificFollowPerson;
